# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkResumeIn": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BulkResumeItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BulkResumeIn",
        "type": "object"
      },
      "BulkResumeItem": {
        "description": "One file in a bulk upload \u2014 creates a new candidate per file.",
        "properties": {
          "content_b64": {
            "title": "Content B64",
            "type": "string"
          },
          "content_type": {
            "default": "application/pdf",
            "title": "Content Type",
            "type": "string"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          }
        },
        "required": [
          "content_b64"
        ],
        "title": "BulkResumeItem",
        "type": "object"
      },
      "BulkResumeOut": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BulkResumeResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "created",
          "results"
        ],
        "title": "BulkResumeOut",
        "type": "object"
      },
...
      "CandidateCreate": {
        "description": "Create a candidate. ``pii`` (name/email/phone/\u2026) is envelope-encrypted at rest.",
        "properties": {
          "consent_state": {
            "default": "pending",
            "title": "Consent State",
            "type": "string"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "pii": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Pii",
            "type": "object"
          }
        },
        "title": "CandidateCreate",
        "type": "object"
      },
      "CandidateDetail": {
        "description": "Full candidate view for the owning org's recruiter (includes decrypted contact).",
        "properties": {
          "consent_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Source"
          },
          "consent_state": {
            "title": "Consent State",
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "resumes": {
            "items": {
              "$ref": "#/components/schemas/ResumeInfo"
            },
            "title": "Resumes",
            "type": "array"
          },
          "review_flags": {
            "items": {
              "$ref": "#/components/schemas/ReviewFlagOut"
            },
            "title": "Review Flags",
            "type": "array"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "external_ref",
          "status",
          "consent_state",
          "consent_source",
          "created_at",
          "contact",
          "skills",
          "experience_years",
          "resumes",
          "review_flags"
        ],
        "title": "CandidateDetail",
        "type": "object"
      },
      "CandidateSummary": {
        "description": "Redacted list/row view \u2014 no raw PII; capability + status only.",
        "properties": {
          "consent_state": {
            "title": "Consent State",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "review_flag_count": {
            "title": "Review Flag Count",
            "type": "integer"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "external_ref",
          "status",
          "consent_state",
          "skills",
          "experience_years",
          "review_flag_count",
          "created_at"
        ],
        "title": "CandidateSummary",
        "type": "object"
      },
      "ConsentEventIn": {
        "description": "Record a consent event (append-only ledger + candidate state).",
        "properties": {
          "event": {
            "title": "Event",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "event"
        ],
        "title": "ConsentEventIn",
        "type": "object"
      },
      "ContactInfo": {
        "description": "Decrypted contact PII \u2014 owning-org detail view only.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "ContactInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeInfo": {
        "description": "One immutable r\u00e9sum\u00e9 version.",
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "parsed": {
            "title": "Parsed",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "content_hash",
          "content_type",
          "is_current",
          "parsed",
          "created_at"
        ],
        "title": "ResumeInfo",
        "type": "object"
      },
      "ResumeUploadIn": {
        "description": "Upload one r\u00e9sum\u00e9 (base64 so no multipart dependency).",
        "properties": {
          "content_b64": {
            "title": "Content B64",
            "type": "string"
          },
          "content_type": {
            "default": "application/pdf",
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          }
        },
        "required": [
          "content_b64"
        ],
        "title": "ResumeUploadIn",
        "type": "object"
      },
      "ResumeUploadOut": {
        "description": "Result of ingesting + parsing one r\u00e9sum\u00e9.",
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "parse_run_id": {
            "title": "Parse Run Id",
            "type": "string"
          },
          "parsed": {
            "title": "Parsed",
            "type": "boolean"
          },
          "resume_id": {
            "title": "Resume Id",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "resume_id",
          "parse_run_id",
          "parsed"
        ],
        "title": "ResumeUploadOut",
        "type": "object"
      },
      "ReviewFlagOut": {
        "description": "An advisory, data-quality flag (never auto-rejects; not fraud detection).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "ReviewFlagOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal AI/services API. Phase 1 product surface over the Phase 0 spine.",
    "title": "ManFriday Internal API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/candidates": {
      "get": {
        "description": "List the org's candidates (redacted \u2014 capability + status, never raw PII).",
        "operationId": "listCandidates",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CandidateSummary"
                  },
                  "title": "Response Listcandidates",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Candidates",
        "tags": [
          "candidates"
        ]
      },
      "post": {
        "description": "Create a candidate; PII (if any) is envelope-encrypted at rest (WP 0.7).",
        "operationId": "createCandidate",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
    "/candidates/bulk": {
      "post": {
        "description": "Bulk import: one new candidate per r\u00e9sum\u00e9 file, each ingested + parsed.",
        "operationId": "bulkUploadResumes",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkResumeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResumeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Upload",
        "tags": [
          "candidates"
        ]
      }
    },
    "/candidates/{candidate_id}": {
      "get": {
        "description": "Full detail for the owning org's recruiter (includes decrypted contact PII).",
        "operationId": "getCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
    "/candidates/{candidate_id}/consent": {
      "post": {
        "description": "Record a consent event: append to the immutable ledger + update candidate state.",
        "operationId": "recordConsent",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Consent",
        "tags": [
          "candidates"
        ]
      }
    },
    "/candidates/{candidate_id}/resumes": {
      "post": {
        "description": "Upload + parse one r\u00e9sum\u00e9 for an existing candidate (immutable store + provenance).",
        "operationId": "uploadResume",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeUploadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeUploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Resume",
        "tags": [
          "candidates"
        ]
      }
    },
...
  }
}